Can now mount cdroms, reading still fails
git-svn-id: svn+ssh://svn.code.sf.net/p/jnode/svn/trunk@342 407389f7-7c16-0410-9eee-ad85acb55d7f

+++ fs/src/fs/org/jnode/fs/ntfs/NTFSVolume.java
@@ -131,7 +131,7 @@ public class NTFSVolume
 		NTFSNonResidentAttribute dataAttribute = (NTFSNonResidentAttribute) this.getMFTRecord().getAttribute(NTFSFileRecord.$DATA);
 		// find out the VCN
 		
-		int offset = this.getBootRecord().getBytesPerFileRecord() * (indexEntry.getFileReferenceNumber());
+		long offset = this.getBootRecord().getBytesPerFileRecord() * (indexEntry.getFileReferenceNumber());
 		
 		// read the buffer
 		byte [] buffer = null;
@@ -145,7 +145,7 @@ public class NTFSVolume
 					this,
 					NTFSUTIL.extractSubBuffer(
 								buffer,
-								offset % this.getClusterSize(),
+								(int)(offset % this.getClusterSize()),
 								this.getBootRecord().getBytesPerFileRecord())
 			);
